// backend/src/common/enums.ts
export enum UserStatus {
  NEEDS_ONBOARDING = 'NEEDS_ONBOARDING',
  ACTIVE = 'ACTIVE',
}

export enum CompanyRole {
  ADMIN = 'ADMIN',
  EMPLOYEE_MANAGER = 'EMPLOYEE_MANAGER',
}

export enum ProjectKind {
  PROJECT = 'PROJECT',
  SUBPROJECT = 'SUBPROJECT',
  ACTIVITY = 'ACTIVITY',
}

export enum Locale {
  EN = 'en',
  DE = 'de',
  FR = 'fr',
}